==> 03.5_Clean_Data.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import Imputer

def findSimilar(ratings, matrix, index):
    npa = np.array(ratings) * np.array(matrix.iloc[:,2:])
    data = pd.DataFrame(npa, index=matrix[index], columns=matrix.iloc[:,2:].columns)

    #display(data.head(15))
    ids = pd.unique(matrix[index])
    packed = pd.DataFrame(index=ids,columns=data.columns)
    #display(packed.head(5))
    #display(packed.loc[272])
    summed = data.groupby([data.index]).max()
    
    # This Imputer is pretty naive. A KNN imputer would be much better.
    impute = Imputer(missing_values='NaN', strategy='mean', axis=0)
    X_imputed = pd.DataFrame(data=impute.fit_transform(summed.as_matrix()))
    X_imputed.index = summed.index
    X_imputed.columns = summed.columns
    display(X_imputed)
    return X_imputed

rawdata = pd.read_csv('boardgame-elite-users.csv')
rawdata=rawdata.rename(columns = {"Compiled from boardgamegeek.com by Matt Borthwick":'userID'})

details = pd.read_csv('boardgame-details.csv')
details = details.rename(columns = {"boardgamegeek.com game ID": 'gameID'})
titles = details.iloc[:,0:2]

gamedata = pd.merge(rawdata.copy(), titles, on='gameID')
X = pd.get_dummies(gamedata, columns=['title'])
X = X.replace(0, np.nan)
del X['gameID']
ratings = X.iloc[:,1:2].copy()

gsim = findSimilar(ratings, X,'userID')
display(gsim.head(6))
gsim.to_csv('gsim.csv', sep='\t')

usim = gsim.transpose()
display(usim.head(6))
usim.to_csv('usim.csv', sep='\t')
