==> 04.1 Collaborative Filter for export.ipynb <==
%matplotlib inline

import pandas as pd
import numpy as np
from scipy.spatial.distance import cosine
from sklearn.preprocessing import Imputer
from sklearn.model_selection import train_test_split

def getSummaries(ratings, matrix, index, impute=True):
    data = pd.DataFrame(np.array(ratings) * np.array(matrix.iloc[:,2:])
        , index=matrix[index], columns=matrix.iloc[:,2:].columns)

    #display(data.head(15))
    ids = pd.unique(matrix[index])
    packed = pd.DataFrame(index=ids,columns=data.columns)
    #display(packed.head(5))
    #display(packed.loc[272])
    summed = data.groupby([data.index]).max()
    
    if impute:
        # This Imputer is pretty naive. A KNN imputer would be much better.
        imp = Imputer(missing_values='NaN', strategy='mean', axis=0)
        X_imputed = pd.DataFrame(data=imp.fit_transform(summed.as_matrix()))
        X_imputed.index = summed.index
        X_imputed.columns = summed.columns
        display(X_imputed.head(5))
        return X_imputed
    return summed

def findNeighbors(data, data_ibs):
    # Lets fill in those empty spaces with cosine similarities
    # Loop through the columns
    for i in range(0,len(data_ibs.columns)) :
        # Loop through the columns for each column
        if i % 100 == 0:
            display(i)
        for j in range(0,len(data_ibs.columns)) :
            # Fill in placeholder with cosine similarities
            data_ibs.iloc[i,j] = 1-cosine(data.iloc[:,i],data.iloc[:,j])
        
    #display(data_ibs.head(5))

    # Create a placeholder items for closes neighbours to an item
    data_neighbours = pd.DataFrame(index=data_ibs.columns,columns=range(1,21))
 
    # Loop through our similarity dataframe and fill in neighbouring item names
    for i in range(0,len(data_ibs.columns)):
        data_neighbours.iloc[i,:20] = data_ibs.iloc[0:,i].sort_values(ascending=False)[:20].index
 
    return data_neighbours


X = pd.read_csv('boardgame-frequent-users.csv')
#X = pd.read_csv('boardgame-elite-users.csv')
X = X.rename(columns = {"Compiled from boardgamegeek.com by Matt Borthwick":'userID'})

X, X_test, y_train, y_test = train_test_split(X, X['rating'], test_size=0.3, random_state=0)

pd.DataFrame(X_test).to_csv("test_data.csv", sep='\t')

X = pd.get_dummies(X, columns=['gameID'])
X = X.replace(0, np.nan)
ratings = X.iloc[:,1:2].copy()

gsum = getSummaries(ratings, X,'userID')
gsum2 = getSummaries(ratings, X,'userID', impute=False)
display(gsum2.head(6))

gsum.to_csv('gsum.csv', sep='\t')
gsum.transpose().to_csv('usum.csv', sep='\t')

gsum2.to_csv('gsum2.csv', sep='\t')
gsum2.transpose().to_csv('usum2.csv', sep='\t')


data_ibs = pd.DataFrame(index=gsum.columns,columns=gsum.columns)
gneighbors = findNeighbors(gsum, data_ibs)
display(gneighbors.head(13))

gneighbors.to_csv('gneighbors.csv', sep='\t')


usum = gsum.transpose()
display(usum.head(6))

data_ibs = pd.DataFrame(index=usum.columns,columns=usum.columns)
uneighbors = findNeighbors(usum, data_ibs)
display(uneighbors.head(6))
uneighbors.to_csv('uneighbors.csv', sep='\t')
