==> 05 Baseline Prediction.ipynb <==
# Baseline naive prediction
#
# Return the mean score for a given game.

import numpy as np
import pandas as pd

rawdata = pd.read_csv('boardgame-elite-users.csv')
rawdata=rawdata.rename(columns = {"Compiled from boardgamegeek.com by Matt Borthwick":'userID'})

details = pd.read_csv('boardgame-details.csv')
details = details.rename(columns = {"boardgamegeek.com game ID": 'gameID'})
titles = details.iloc[:,0:2]
titles = titles.set_index('gameID')

gamedata = rawdata.copy()
gamedata = gamedata.groupby(['gameID'])['rating'].mean()

def predict(titles, gamedata, gameID):
    display(titles.loc[gameID])
    display(gamedata[gameID])

predict(titles, gamedata, 150)
predict(titles, gamedata, 13)
predict(titles, gamedata, 18)
predict(titles, gamedata, 181)
predict(titles, gamedata, 77130)
predict(titles, gamedata, 98)
