==> 06 Local test.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error
from sklearn.metrics import median_absolute_error
from math import sqrt

usum = pd.read_csv('usum2.csv', sep='\t', index_col=0)
gsum = pd.read_csv('gsum2.csv', sep='\t', index_col='userID')
#unbr = pd.read_csv('uneighbors.csv', sep='\t', index_col=0)
#gnbr = pd.read_csv('gneighbors.csv', sep='\t', index_col=0)
#del gnbr['1']
#del unbr['1']
        
# user neighbors mean scores for the game
# averaged with
# game neighbors mean scores for the user
def predict(gameID, userID):
    global usum, gsum
    #, unbr, gnbr
    # Series of games in rank - neighbors of original
    #display(gnbr.loc['gameID_'+gameID])
    # All the scores for one game
    #display(usum.loc['gameID_'+gameID])
    # Series of users in rank - neighbors of original
    #display(unbr.loc[userID])
    # All the ratings from one user
    #display(gsum.loc[userID])
    #gameList = gnbr.loc[['gameID_'+gameID]].transpose()
    #gameList = gameList[gameList.columns[0]].tolist()
    #display(gnbr.head(3))
    #display(gsum.head(5))
    #display(gsum.loc[[userID]])
    
    # median of user scores for neighboring games
    gavg = gsum.loc[[userID]].transpose().dropna().median()
    #display(gavg)
    
    #userList = unbr.loc[[userID]].transpose()
    #userList = list(map(str, userList[userList.columns[0]].tolist()))
    # userList are numbers, but need string references for column list
    #display(userList)
    #display(usum.loc[['gameID_'+gameID]][userList])
    # median of neighboring users scores for this game
    uavg = usum.loc[['gameID_'+gameID]].transpose().dropna().median()

    # TODO - use a better filler value than avg
    if uavg.isnull().any():
        #uavg.values[0] = 7.2
        uavg = usum.loc[['gameID_'+gameID]].transpose().dropna().median()
    if gavg.isnull().any():
        #gavg.values[0] = 7.2
        gavg = gsum.loc[[userID]].transpose().dropna().median()
    return (uavg.values[0]+gavg.values[0])/2

#display(gsum.index)

#predict('13', 272)
test = pd.read_csv('test_data.csv', sep='\t')
y_actual = test['rating']
y_predicted = []

for index, row in test.iterrows():
    gid = str(int(row.values[2]))
    uid = int(row.values[1])
    if uid in gsum.index:
        p = predict(gid, uid)
        #display(p)
        y_predicted.append(p)

rms = sqrt(mean_squared_error(y_actual, y_predicted))
display(rms)
display(median_absolute_error(y_actual, y_predicted))
display(mean_absolute_error(y_actual, y_predicted))
#test.to_csv('filled-test-tgreiser.csv', sep='\t')

# median for NaN values
# 1.438198 / 0.75 / 1.05731
# 7.2 for NaN values
# 1.452891 / 0.75 / 1.062215
# median for NaN, mean for neighbors
# 1.41465 / 0.78 / 1.04765
# user mean only
# 1.5523 / 0.848 / 1.43
# game mean only
# 1.44245 / 0.8 / 1.063789
#
# frequent users
# all games mean, ignore neighbors
# 1.3389 / 0.77 / 0.997
# all games median, ignore neighbors
# 1.3528 / 0.75 / 0.988
