==> 06 Surprise 4.ipynb <==
from surprise import *

from surprise import Dataset
from surprise import evaluate, print_perf
from surprise import Reader
import pandas as pd
from os import path

train_file = path.expanduser('./full-train.csv')
test_file = path.expanduser('./full-test.csv')

#reader = Reader(line_format='userID gameID rating', sep=',', rating_scale=(1, 10))
reader = Reader(line_format='user item rating', sep=',', skip_lines=1, rating_scale=(1, 10))
folds_files = [(train_file % i, test_file % i) for i in (1, 2, 3, 4, 5, 6, 7, 8 , 9, 10)]
data = Dataset.load_from_folds(folds_files, reader=reader)

#SVD - param_grid = {'n_epochs': [10], 'reg_all': [0.02], 'biased': [True]}
# 1.1889 / 0.7134

param_grid = {'n_epochs': [10], 'reg_all': [0.02], 'lr_all': [0.05, 0.07]}
grid_search = GridSearch(SVDpp, param_grid, measures=['RMSE', 'FCP'],
                         verbose=True)
grid_search.evaluate(data)

#1.29126127498
#{'bsl_options': {'method': 'als', 'reg': 1}, 'k': 3, 'sim_options': {'name': 'msd', 'min_support': 1, 'user_based': False}}
#0.680624884689
#{'bsl_options': {'method': 'als', 'reg': 1}, 'k': 3, 'sim_options': {'name': 'msd', 'min_support': 1, 'user_based': False}}

# best RMSE score
print(grid_search.best_score['RMSE'])
# >>> 0.96117566386

# combination of parameters that gave the best RMSE score
print(grid_search.best_params['RMSE'])
# >>> {'reg_all': 0.4, 'lr_all': 0.005, 'n_epochs': 10}

# best FCP score
print(grid_search.best_score['FCP'])
# >>> 0.702279736531

# combination of parameters that gave the best FCP score
print(grid_search.best_params['FCP'])
