==> 06 Surprise Prediction.ipynb <==
from surprise import *

from surprise import Dataset
from surprise import evaluate, print_perf
from surprise import Reader
import pandas as pd
from os import path

file_path = path.expanduser('./boardgame-frequent-users.csv')

#reader = Reader(line_format='userID gameID rating', sep=',', rating_scale=(1, 10))
reader = Reader(line_format='user item rating', sep=',', skip_lines=1, rating_scale=(1, 10))
data = Dataset.load_from_file(file_path, reader=reader)
trainset = data.build_full_trainset()

algo = SVD(n_epochs=10, reg_all=0.02, n_factors=100)
algo.train(trainset)
#pred = algo.predict(str(13), str(272))
#display(pred.est)
# 6.881262849120394

gsum = pd.read_csv('gsum.csv', sep='\t', index_col='userID')
test = pd.read_csv('boardgame-users-test.csv', sep=',')
display(test.shape)
display(test.head(5))
for index, row in test.iterrows():
    gid = str(int(row.values[1]))
    uid = str(int(row.values[0]))
    rating = algo.predict(uid, gid)
    #display(rating)
    if rating.details['was_impossible'] == False and int(uid) in gsum.index:
        #display(index)
        test.iloc[index:index+1,2:3] = rating.est
    
test.to_csv('filled-svd-tgreiser.csv', sep=',')
display('done')
