==> 07 User counts.ipynb <==
import pandas as pd
from random import random, seed


display('Begin')
seed()
X = pd.read_csv('boardgame-users.csv')
X = X.rename(columns = {"Compiled from boardgamegeek.com by Matt Borthwick":'userID'})
X.set_index('userID')
X = X.sample(frac=1)

counts = X.groupby(['userID']).size().sort_values()
display(counts.head(5))

counts.to_csv('user_counts.csv')

total = X.shape[0]
testCount = int(total * .3)

display(testCount)

train = []
test = []

for index, row in X.iterrows():
    uid = int(row.values[0])
    prob = 0.0
    if counts[uid] > 1:
        prob = 0.1
    elif counts[uid] > 2:
        prob = 0.2
    elif counts[uid] > 3:
        prob = 0.3
    elif counts[uid] > 4:
        prob = 0.4
    elif counts[uid] > 5:
        prob = 0.6
    elif counts[uid] > 6:
        prob = 0.8
    elif counts[uid] > 7:
        prob = 0.9
    elif counts[uid] > 8:
        prob = 0.95
    
    if len(test) < testCount and random() < prob:
        test.append(row)
        counts[uid] -= 1
    else:
        train.append(row)

display(len(train))
display(len(test))
display(len(test)/len(train))
pd.DataFrame(train).to_csv('full-train.csv')
pd.DataFrame(test).to_csv('full-test.csv')
